Stable build from zip archive release + patch
The included Makefile for building the engine uses the wrong variable
name (FTEQCC should be QCC) so this is patched during build time.

Note that despite using the release archive this does build the actual
game engine for the enviroment. The game asset files (pk3) are the pre-
built releases from the archive.